Find a way to refactor a certain block of code and turn it into more clean code.
Actually I have another solution for this problem: We can do an array of object so each object in array includes all data of one project such as name, link, image , etc..
and then display it with "innerHTML" using for loop.
I'll be glad to listen to your opinion about this solution.